ai-service/requirements.txt
fastapi
uvicorn[standard]
mediapipe
opencv-python
numpy
python-multipart
python-dotenv